feat(physical_plan): spillable sinks — grace hash join, external sort, grace aggregate
Add disk-backed ("spilling") variants of the three blocking physical
operators so large joins, sorts and aggregations no longer have to fit
their full build/sort/group state in memory.

Operators (standalone, no shared base with the in-memory ones — the
in-memory operator_{hash_join,sort,group} stay byte-identical to main):
  - operator_grace_hash_join_t: partitioned (grace) hash join; both
    inputs are hash-partitioned to disk, then each partition pair is
    joined in memory.
  - operator_external_sort_t: external merge sort; sorted runs are
    spilled and k-way merged.
  - operator_external_group_t: two-phase partitioned grace aggregate;
    groups are hash-partitioned to disk and merged per partition with a
    commutative, count-weighted-mean recombination.

Lowering is proactive (R3/R6): the optimizer rule spill_strategy stamps
exec_strategy=spill on the logical node and the physical-plan generator
lowers to the spilling operator. The operators never inspect memory at
runtime and never fall back — the annotation is the single source of
truth. Default in_memory preserves pre-spill behaviour. Spill is opt-in
via configuration (config.disk.spill_enabled, default false).

On-disk format: a self-describing, length-prefixed Unified Format
(OTSC1.0) serializer (serialize_unified / deserialize_unified) carrying
the MVCC visibility header and row-group layout; nested LIST/ARRAY/STRUCT
are walked via the child vectors directly (no logical_value_t round-trip,
R1). I/O goes through core::filesystem; spill_file_t is a thin pure-value
handle that always cleans up its run files.

AVG fix: avg_finalize now finalizes to DOUBLE across the numeric and
DECIMAL domains (scale-aware mantissa / 10^scale), with HUGEINT/UHUGEINT
handled defensively and an explicit conversion_failure on unknown types
instead of a silent garbage reinterpret. Adds numeric_or_decimal matcher
so AVG(DECIMAL) resolves a kernel.

Error handling follows docs/error_handling.md: serializer entry points
return core::error_t / core::result_wrapper_t (no bool-ok, no throw on
the hot or cross-actor paths).

Tests: grace-join / external-sort / grace-aggregate spill paths
(test_spill_red, test_sort_spill), unified-format round-trip incl.
nested types and scaled DECIMAL, and AVG(DECIMAL) scale-awareness.
Full suite: 964/964 green.

+++ integration/cpp/test/test_arithmetic.cpp
@@ -376,10 +376,10 @@ TEST_CASE("integration::cpp::test_arithmetic") {
                                            R"_(FROM TestDatabase.TestCollection;)_");
         REQUIRE(cur->is_success());
         REQUIRE(cur->size() == 1);
-        // avg(1..100) = 50.5, val = 50.5 * 10 = 505
-        // AVG might return int or double depending on implementation
-        auto val = cur->chunks().front().data[0].data<int64_t>()[0];
-        REQUIRE(val == 505);
+        // avg(1..100) = 50.5, val = 50.5 * 10 = 505.0. AVG now always returns a
+        // DOUBLE (avg_finalize), so read the column as double, not int64.
+        auto val = cur->chunks().front().data[0].data<double>()[0];
+        REQUIRE(val == 505.0);
     }
 
     INFO("C4. MIN/MAX of expression") {
